fix convo operator add batch norm max pool operators

+++ rohithk/deeplearning_operators_assignment/Operators/include/ctensor.h
@@ -7,10 +7,10 @@
 class CTensor {
 public:
   std::vector<int> shape;
-  std::vector<float> data;
+  std::vector<double> data;
 
   CTensor();
-  CTensor(const std::vector<int> &shape_, float val = 0.0);
+  CTensor(const std::vector<int> &shape_, double val = 0.0);
   CTensor(const CTensor &copy);
 };
